requirements.txt
Django>=4.2,<5.0
whitenoise>=6.6.0
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
pyinstaller>=6.0.0
requests>=2.31.0
